Extract the survey-rating lookup out of the chat-history cell
Moves the message-finding/id-parsing/metadata-lookup logic for a
conversation's zd:surveys rating into getZendeskSurveyRating() in
utils.tsx, alongside the sibling helpers that already scan a
conversation's messages the same way (e.g. getClientId). Keeps the
component body focused on rendering.
